# Further improve natural sorting of classes