--- a/packages/tailwindcss/src/utils/compare.ts
+++ b/packages/tailwindcss/src/utils/compare.ts
@@ -20,9 +20,9 @@ export function compare(a: string, z: string) {
     // If both are numbers, compare them as numbers instead of strings.
     if (aCode >= ZERO && aCode <= NINE && zCode >= ZERO && zCode <= NINE) {
       let aStart = i
-      let aEnd = i
+      let aEnd = i + 1
       let zStart = i
-      let zEnd = i
+      let zEnd = i + 1
 
       // Consume the number
       while (a.charCodeAt(aEnd) >= ZERO && a.charCodeAt(aEnd) <= NINE) aEnd++