This PR is a small continuation of #13507 where @rvanvelzen pointed out that we are checking a character that we already compared.

Additionally, this PR also prevents re-computing the `charCode` twice.

Performance wise, this is very minimal, running this on the tailwindcss.com codebase I get these results:

Before:
```
[3.19ms] Sort
[3.06ms] Sort
[3.13ms] Sort
[3.38ms] Sort
[3.11ms] Sort
[3.05ms] Sort
[3.03ms] Sort
[3.02ms] Sort
[2.96ms] Sort
[3.18ms] Sort
-------------
[3.171ms] avg
```
After:
```
[3.24ms] Sort
[3.02ms] Sort
[3.14ms] Sort
[3.05ms] Sort
[3.17ms] Sort
[3.09ms] Sort
[3.02ms] Sort
[3.12ms] Sort
[2.94ms] Sort
[2.97ms] Sort
-------------
[3.076ms] avg
```

<!--

👋 Hey, thanks for your interest in contributing to Tailwind!

**Please ask first before starting work on any significant new features.**

It's never a fun experience to have your pull request declined after investing a lot of time and effort into a new feature. To avoid this from happening, we request that contributors create an issue to first discuss any significant new features. This includes things like adding new utilities, creating new at-rules, or adding new component examples to the documentation.

https://github.com/tailwindcss/tailwindcss/blob/master/.github/CONTRIBUTING.md

-->